Major_Bug - added extra security check for required Category based Api methods to prevent users to modify other user's category data

+++ src/main/java/com/manga/collectionBend/service/CategoryService.java
@@ -62,9 +62,7 @@ public class CategoryService {
                 .toList();
     }
 
-    public ApiResponse<CategoryResponse> addCategoryHandler(CategoryRequest categoryRequest) {
-//        get userId from requestBody from frontend
-        Integer userId = categoryRequest.getUserId();
+    public ApiResponse<CategoryResponse> addCategoryHandler(CategoryRequest categoryRequest, Integer userId) {
 //        based on userId value- get UserEntity reference data to link it/store it in Category table-column
         UserEntity user = userRepo.findById(userId).orElse(null);
 
@@ -95,14 +93,14 @@ public class CategoryService {
         return ApiResponse.success(dto);
     }
 
-    public CategoryResponse updateCategoryHandler(Integer categoryId, CategoryRequest categoryRequest) {
+    public CategoryResponse updateCategoryHandler(Integer categoryId, CategoryRequest categoryRequest, Integer userId) {
 //       Validation check - get Category data based on provided category-id
         CategoryEntity existingCategory = categoryRepo.findById(categoryId)
                 .orElseThrow(() -> new RuntimeException("Category not found with id = " + categoryId));
 
 //        UserId Validation check- to see if same user is trying to update his data or some one
 //        if same - we update or not same - we throw error
-        if(Objects.equals(categoryRequest.getUserId(), existingCategory.getUser().getUserId())) {
+        if(Objects.equals(userId, existingCategory.getUser().getUserId())) {
             // mutate the existing managed entity directly — don't rebuild it
 //            setting values to already existing entity object which has updated categoryName value and remaining 2 field - id and User values will be same as previous/existing category record-data
             existingCategory.setCategoryName(categoryRequest.getCategoryName()); // new updated value
@@ -118,45 +116,49 @@ public class CategoryService {
             dto.setCategoryName(updatedCategory.getCategoryName());
             return dto;
         } else {
-            System.out.println("Error - provided UserId is not matching with existing category Id, hence not updating any data!!!");
-            return new CategoryResponse();
+            throw new IllegalStateException("You userId: "+ userId +" are not authorized to update other user's data!");
         }
     }
 
-    public CategoryDeleteResponse deleteCategoryHandler(Integer categoryId) {
+    public CategoryDeleteResponse deleteCategoryHandler(Integer categoryId, Integer userId) {
         //       Validation check - get Category data based on provided category-id
         CategoryEntity existingCategory = categoryRepo.findById(categoryId)
                 .orElseThrow(() -> new RuntimeException("Category not found with id = " + categoryId));
 
         String categoryName = existingCategory.getCategoryName();
-        int userId = existingCategory.getUser().getUserId();
+        //        UserId Validation check- to see if same user is trying to update his data or some one
+//        if same - we delete or not same - we throw error
+        if(Objects.equals(userId, existingCategory.getUser().getUserId())) {
 //        Logic to prevent Category Deletion if that category has some created collections which are linked/mapped to this wanted category
 //        here when Category-record/row will act like parent record which is referred by foreign-key in a Collection-record/row(child) -
 //        so delection of parent record(category) will not work until its referred child record in Collection table is deleted first then only deletion of category-record will work next
-        List<CollectionEntity> categoryBasedCollections = collectionRepo.findByCategory(existingCategory);
+            List<CollectionEntity> categoryBasedCollections = collectionRepo.findByCategory(existingCategory);
 
 //        delete category based collections images logic to delete each collection image separately-
 //        - because we cant delete entire userId-main folder of User just to delete few collectionImages which are only linked to this CategoryId
-        if (categoryBasedCollections != null && !categoryBasedCollections.isEmpty()) {
-            categoryBasedCollections.forEach(collection -> {
-                //        only delete file/imagge in CollectionImages-backend-folder if in Db it has imageName stored(means user given img while creating this collection) or else imageName="" empty(means user did no give any img which does not need deleting anything)
-                if(collection.getImagename() != null && !Objects.equals(collection.getImagename(), "")){
-                    String newPath = path + File.separator + userId;
-                    //        delete the file/image associated with this object which will be deleted from table/db
-                    try {
-                        Files.deleteIfExists(Paths.get(newPath + File.separator + collection.getImagename())); // deletes empty folder or single file once
-                    } catch (IOException e) {
-                        throw new RuntimeException(e);
+            if (categoryBasedCollections != null && !categoryBasedCollections.isEmpty()) {
+                categoryBasedCollections.forEach(collection -> {
+                    //        only delete file/imagge in CollectionImages-backend-folder if in Db it has imageName stored(means user given img while creating this collection) or else imageName="" empty(means user did no give any img which does not need deleting anything)
+                    if(collection.getImagename() != null && !Objects.equals(collection.getImagename(), "")){
+                        String newPath = path + File.separator + userId;
+                        //        delete the file/image associated with this object which will be deleted from table/db
+                        try {
+                            Files.deleteIfExists(Paths.get(newPath + File.separator + collection.getImagename())); // deletes empty folder or single file once
+                        } catch (IOException e) {
+                            throw new RuntimeException(e);
+                        }
                     }
-                }
-            });
-        }
+                });
+            }
 
 //  Here category record data will get deleted and automatically its linked collection records will also get deleted by Hibernate-Cascade.ALL
-        categoryRepo.delete(existingCategory);
-        CategoryDeleteResponse categoryDeleteResponse = new CategoryDeleteResponse();
-        categoryDeleteResponse.setSuccess(Boolean.TRUE);
-        categoryDeleteResponse.setMessage("Category deleted successfully with name = " + categoryName);
-        return categoryDeleteResponse;
+            categoryRepo.delete(existingCategory);
+            CategoryDeleteResponse categoryDeleteResponse = new CategoryDeleteResponse();
+            categoryDeleteResponse.setSuccess(Boolean.TRUE);
+            categoryDeleteResponse.setMessage("Category deleted successfully with name = " + categoryName);
+            return categoryDeleteResponse;
+        }else {
+            throw new IllegalStateException("You userId: "+ userId +" are not authorized to delete other user's data!");
+        }
     }
 }
